+++ src/UI/Widgets/Interfaces/IWidget.cs
@@ -3,6 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Xna.Framework;
+using WForest.Rendering;
 using WForest.UI.Interactions;
 using WForest.UI.Utils;
 using WForest.Utilities;
@@ -25,6 +26,12 @@ namespace WForest.UI.Widgets.Interfaces
         /// - Released
         /// </summary>
         public Interaction CurrentInteraction { get; set; }
+
+        /// <summary>
+        /// Draw the widget with a SpriteBatch.
+        /// </summary>
+        /// <param name="renderer"></param>
+        void Draw(IRenderer renderer);
 
         #region Widget Tree