+++ CHANGELOG.md
@@ -4,10 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.0.15]
+### Added
+- IRenderer interface to abstract over Spritebatch
+- DefaultRenderer that uses a Spritebatch to draw widgets
+
 ### Changed
 - Constructors of Props and Widgets are now public
 - PropertyFactory in PropFactory
 - Refactored Rectangle usage. The library now uses floats instead of ints through a custom RectangleF struct replacing the monogame int Rectangle
+- Draw method moved into Widget and now takes an IRenderer instead of a Spritebatch
 
 ### Fixed
 - Vertical/Horizontal stretch now respect spaces of sibling and stretch as long as siblings have space